// src/app/shared/swagger/model/stepDto.ts
/**
 * Backend
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface StepDto { 
    id?: number;
    stepNumber?: number;
    checklistId?: number;
    stepDescription?: string | null;
    stepName?: string | null;
}